refactor(beacon): extract beacon parsing internals to BeaconParser.kt (#265)
